test(logger): test_logger.pyを4ファイルに責務分割 (#530)
【背景】test_logger.pyが1530行に肥大化し、Sentry関連39テストが
単一クラスに集中してレビュー・保守が困難だった。

【変更内容】coreロジック5クラスを残置(172行)し、Sentry関連テストを
capture経路/throttle経路/PII検証の3テーマ別ファイルへ純粋移動分割。
重複していたヘルパー_make_fake_validation_error(6箇所)をhelpers.pyへ
統合。各新ファイルにpytestmarkとimportを再宣言。

【影響範囲】挙動変更なし(純粋移動)。全57テストの内容・件数は移動前後で
完全一致。

品質ゲート: 1443 tests passed / 97.58% coverage / ruff 0 errors / mypy 0 errors
